CREATE: navigator package and navigator button
